app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.3
PyYAML>=6.0.1
onnxruntime>=1.17.0
pytest>=8.3.0
httpx>=0.27.0
